Organize adjoint lowering into a coherent module
Relocate the forward pass into the same location as the reverse pass, as
having the two pieces of the pass in different locations is very
confusing. A subdirectory is introduced for the pass for better
organization.
