+++ mlir/lib/Quantum/Transforms/AdjointLowering/QuantumSplitting.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "Quantum/Utils/QuantumSplitting.h"
+#include "QuantumSplitting.hpp"
 
 #include "mlir/Dialect/Complex/IR/Complex.h"
 #include "mlir/Dialect/Func/IR/FuncOps.h"
