+++ mlir/lib/Quantum/Utils/CMakeLists.txt
@@ -1,4 +1,3 @@
 add_mlir_library(QuantumUtils
-	QuantumSplitting.cpp
 	RemoveQuantum.cpp
 )
